backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pillow>=10.0.0
python-multipart>=0.0.20
oagi>=0.12.0
pyautogui>=0.9.54
pydantic>=2.0.0
